Improve the script to create database and user on MySQL.
When creating a user, using 'create user ...', it will fail
if the user already exists:

Error Code: 1396. Operation CREATE USER failed for 'blaine'@'localhost'

The alternative it to use grant - because grant will create the user if the
account does not already exist.

grant seems to be the most elegant way to create a use on MySQL.

http://www.cryer.co.uk/brian/mysql/how_to_create_user_if_not_exist.htm

Signed-off-by: Thiago Farina <tfarina@chromium.org>