network: Initialize |addrlen| variable.
Without initializing |addrlen| to the value of sizeof(struct sockaddr_in), we
were receiving the wrong value from recvfrom because we were not passing
a valid pointer and that was causing sendto to fail.

Answer: http://stackoverflow.com/a/13898228/5237416

By initializing it, the server is now capable to reply to the client
with 'ack 0' without issues and the client does not get stuck anymore.

Signed-off-by: Thiago Farina <tfarina@chromium.org>