go/mysql: use only GRANT to create a user on MySQL
When creating a user, using 'create user ...', it will fail if the user already
exists:

Error Code: 1396. Operation CREATE USER failed for 'nash'@'localhost'

GRANT will create the user if the account does not already exist.

GRANT seems to be the most elegant way to create a user on MySQL.

http://www.cryer.co.uk/brian/mysql/how_to_create_user_if_not_exist.htm

Signed-off-by: Thiago Farina <tfarina@chromium.org>