unixserver.c: add all the debugging code
They were necessary to catch the error "address already in use" in bind,
that was preventing it from sending the message to the client.

The problem is that we have to handle some signal, to do a cleanup,
like removing the server.socket file before exiting the server.

That should prevent the bind() error that was making accept() to return
"invalid argument".

Signed-off-by: Thiago Farina <tfarina@chromium.org>