echo: add file necessary for CMake meta-build system
That way we can build with either Make or Ninja, but more importantly
write code that is portable to other systems, thing that is not possible
directly on Ninja.

With CMake it is also possible to have a different set of flags and
defines depending on the build type (debug vs release).

To build echod you can now do:

$ mkdir -p out
$ cd out
$ cmake -GNinja ..
$ ninja

Signed-off-by: Thiago Farina <tfarina@chromium.org>