bazel/project1: adding dependencies
This creates separate rules for each subcomponent and make them
depend on each other.

This builds the same files as before, but in a different way:
now Bazel will build the "greeter" library first and then build
"my-other-runner".

Now if you edit "ProjectRunner.java" and rebuild "my-other-runner",
"Greeting.java" will not need to be recompiled

Signed-off-by: Thiago Farina <tfarina@chromium.org>