echod: only a negative value is an output error
man 3 snprintf:

RETURN VALUE

Upon successful return, these functions return the number of
characters printed (excluding the null byte used to end output to
strings).

The functions snprintf() and vsnprintf() do not write more than size
bytes (including the terminating null byte ('\0')).  If the output
was truncated due to this limit, then the return value is the number
of characters (excluding the terminating null byte) which would have
been written to the final string if enough space had been available.
Thus, a return value of size or more means that the output was
truncated.  (See also below under NOTES.)

If an output error is encountered, a negative value is returned.

http://man7.org/linux/man-pages/man3/printf.3.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>