csvcat.c: first allocate memory for the rows
The calculation is as follows:

Each row is a pointer to double, so, for example, if we are in a
32-bit machine, the size of 'pointer to double' will be 4 bytes.

Say we have 252 lines (rows) in the csv file we got, skipping
the first line (header).

If we need to allocate memory for that 252 lines then we have:

4 * 252

Which, when translated to C, is:

sizeof(double *) * numrows

That will allocate memory for a 'vector' or 'array', each line
containing an address, pointing to a memory location.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>