csvcat.c: now really allocate memory for each individual row
In order to do this, we need to know the size of 'double', which
is the data type that represents each individual column in the
csv file. Then we need to know how many columns each row has.

With the size of 'double' data type and the number of 'columns',
we can calculate how many bytes we need to allocate for each row.

If we are in a 32-bit machine, the size of double will be 8 bytes, if we have 7 columns in the csv file, then each row needs 56 bytes:

8 * 7 = 56

Translating that to C:

sizeof(double) * state.columns

Signed-off-by: Thiago Farina <tfransosi@gmail.com>