csvcat.c: store volume data in a separate array (or vector)
To calculate the memory necessary to be allocated, we have to
compute the size of 'int' and multiply that by the number of rows (lines).

In C that translates to:

sizeof(int) * numrows

So if we are in a 32-bit machine, the size of 'int' data type will be
4 bytes (32 bits) and if we have 252 lines, then:

4 * 252 = 1008 (bytes)

will be allocated.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>