vector: for the user to tell 'vector_t' the size of the elements it will hold
This is because in in C, you can create an array through two ways.

The first method is declaring it on the stack like so:

int array[50];

The second method is declaring a pointer then calling malloc (on the heap):

int *array;

array = malloc(sizeof(int) * 50);

What the malloc paramenter above does? It informs malloc that it whats
it to allocate memory for 50 elements of the size of a integer, which is
4 bytes or 32 bits in a 32-bit machine. So it is asking malloc to allocate
a region with 200 bytes and return the address to the start of that region.

If you don't multiple 50 by sizeof(int) you can't have enough memory to
store each integer you want in the contigous memory space, because you need
4 bytes to store one integer in memory. That is why you have to multiply by
sizeof(int).

Signed-off-by: Thiago Farina <tfransosi@gmail.com>