vector: assign the new element using pointer arithmetic syntax
The thing is, arr[i] is defined using pointer arithmetic, so:

arr[i] = *(arr + i)

The truth is: arr is a pointer to arr[0].

In fact, arr is defined to be:

& arr[ 0 ]

What does that mean? It means that arr (the variable name) is a POINTER to the first element in the array.

So if you do arr[i], then in really the compiler translates that to (arr + i), and it is just a pointer.

So if you want to assign the value at that location you use the dereference operator ('*') and you are done.

https://www.cs.swarthmore.edu/~richardw/classes/cs31/s18/offsite/pointer.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>