echod: point the user to the right direction
Explain to the user his mistake, so he can correct it right way. A simpler error
message should help on that.

Also stop allowing 0 as port number. 0 is reserved by IANA for both TCP and UDP protocols.

https://en.wikipedia.org/wiki/List_of_TCP_and_UDP_port_numbers

Tested with the following command lines:

$ sudo ./build/echod -f -p 0
$ sudo ./build/echod -f -p 65536

As expected both report the error message for the invalid port number.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>