echod: change the pid file owner, to be able to remove it later
That does not fix the issue unfortunately due to how permissions work on linux.

error: unable to remove the PID file '/var/run/echod.pid': Permission denied

According to the answer in
https://unix.stackexchange.com/questions/49638/cannot-delete-a-file-permission-denied-why:

"In order to delete a file, you must have write permissions on the directory that the file resides in."

So there you have your answer. Even owning the file by the user does not allow it to remove from /var/run,
because 'run' directory is owned by root.

In order to fix that, what we need to do is to create a new directory under /var/run. The only example I could
find that does this was avahi-daemon.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>