echod: create the directory where we will put the pid file
This should fix the following error, that we are trying to fix a long time:

error: unable to remove the PID file '/var/run/echod.pid': Permission denied

The reason this error occurs is that [1]:
"In order to delete a file, you must have write permissions on the directory that the file resides in."

[1] https://unix.stackexchange.com/questions/49638/cannot-delete-a-file-permission-denied-why

So creating a new directory under /var/run and owning it, should allow us to remove the files in it.

Thanks to pacemaker.
https://github.com/ClusterLabs/pacemaker/blob/60504e0676245da3a9910c8b50171c5b472e7749/daemons/pacemakerd/pacemakerd.c#L1156

Signed-off-by: Thiago Farina <tfransosi@gmail.com>