gensignals.c: use %*d notation to align the output of order rows too
The given integer between the % and format specifier characters denote the
number of spaces to be filled on the left side of the output item.
Hence the output now looks like this:

Order
    1
    2
    3
    .
    .
    .

The %*d notation additionally requires the integer to be supplied as argument,
hence we pass ordercolw to printf as argument as well.

https://www.delftstack.com/howto/c/printf-align-columns-in-c/

Signed-off-by: Thiago Farina <tfransosi@gmail.com>