gensignals.c: compact the output of "Signal" column
And align the strings "BUY" and "SELL" on the right.

The number six (6) in %6s, was chosen because it is the length of "Signal" string,
and bigger than four for "SELL" and three for "BUY", so in order to align right,
that tell printf to fill the left side with empty spaces.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>