gensignals.c: compact the output of "P/L" column
Since it is the last column it is not necessary to say the number of characters
the column must have, that is why the format is string for it is only %s.

For the data rows when printing the values, the format string is %-9.2f, to
align on the left (the minus sign -), nine (9) is to leave a big room for big numbers,
and two (2) is because we want to see only two decimal places.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>