pidfile.c: place the name of the function in column one
According to GNU Coding Standards:
5.1 Formatting Your Source Code
"It is important to put the open-brace that starts the body
of a C function in column one, so that they will start a defun.
Several tools look for open-braces in column one to find the
beginnings of C functions.
It is also important for function definitions to start the name
of the function in column one. This helps people to search for
function definitions, and may also help certain tools recognize them."

https://www.gnu.org/prep/standards/html_node/Formatting.html

Function Declarations: https://git.sr.ht/~sircmpwn/cstyle#function-declarations
Placing the name at the start of the line makes it easier to find
functions in a large codebase with e.g.:

git grep '^my_function'

Signed-off-by: Thiago Farina <tfransosi@gmail.com>