echod: drop anciliary groups first
The following is the full explanation given in by O'REILLY from Secure Programming Cookbook for C and C++.

"Ancillary groups are inherited by a process from its parent process, and
they can only be altered by a process with superuser privileges. Therefore,
if a process with superuser privileges is dropping these privileges, it must
also be sure to drop any ancillary groups it may have. This is achieved by calling
setgroups( ) with a single group, which is the real group ID for the process.
Because the setgroups( ) system call is guarded by requiring the effective user ID
of the process to be that of the superuser, it must be done prior to dropping root privileges."

https://www.oreilly.com/library/view/secure-programming-cookbook/0596003943/ch01s03.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>