vector: call CMake's enable_testing() function
It enables testing for current directory and below. It should be called
in the source directory root because ctest expects to find a test file
in the build directory root.

When generating the project files, CMake will create a CTestTestfile.cmake
in the build directory that ctest expects to be there.

https://cmake.org/cmake/help/latest/command/enable_testing.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>