echoc: include the file name after the @file command
Doxygen says "If the file name is omitted (i.e. the line after \file is left blank)
then the documentation block that contains the \file command will belong to the file
it is located in."

Since the line after @file is not blank, let's put the file name for it.

https://www.doxygen.nl/manual/commands.html#cmdfile

Signed-off-by: Thiago Farina <tfransosi@gmail.com>