sab: do not give "Ctrl+D" as shortcut for Delete
This simplifies the usuability for the users.

From Microsoft UX Guide Lines:

Windows 7 Menus (Design basics)

"Standard menu bars

The standard menu bar structure is as follows. This list shows the menu category and item labels, their order with separators, their access and shortcut keys, and their ellipses."

Delete Del (shortcut usually not given)

https://learn.microsoft.com/en-us/windows/win32/uxguide/cmd-menus#access-keys

Since the mnemonic is under the letter 'D', that already gives a hint to the
user of which key he can use to delete an item. It can be either 'd' or Del.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>