sab: cleanup the code in '_execute_sql' function
Use sqlite3_exec which basically does pretty much the same, but
being a wrapper around these other sqlite3 functions, it helps
simplify the code. And less code, means it will be easier when
porting this code to Windows.

So with just sqlite3_exec, I can test things on Windows much faster
and with less error and debugging.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>