sab: remove integer literals from sql insert statement
We don't need to use integer literals in the sql insert statement
parameters.

According to the SQLite documentation, the parameter can be just a
question mark (?), also the leftmost SQL parameter has an index of 1,
so it knows it starts at 1.

" In the SQL statement text input to sqlite3_prepare_v2() and its variants, literals may be replaced by a parameter that matches one of following templates:

    ?
    ?NNN
    :VVV
    @VVV
    $VVV

The second argument is the index of the SQL parameter to be set. The leftmost SQL parameter has an index of 1.
"

https://www.sqlite.org/c3ref/bind_blob.html

https://stackoverflow.com/questions/19614544/sql-question-marks-in-insert-statement
https://stackoverflow.com/questions/3727688/what-does-a-question-mark-represent-in-sql-queries

Signed-off-by: Thiago Farina <tfransosi@gmail.com>