WunderPeople: promote 'hdb' variable to file scope
That way it can be accessed by other functions within this file when we start
adding more functions, as we will very soon.

From Microsoft's C Language Reference:

"File scope
The declarator or type specifier for an identifier with file scope appears outside any block or list of parameters and is accessible from any place in the translation unit after its declaration. Identifier names with file scope are often called "global" or "external." The scope of a global identifier begins at the point of its definition or declaration and terminates at the end of the translation unit."

https://learn.microsoft.com/en-us/cpp/c-language/scope-and-visibility?view=msvc-170

Signed-off-by: Thiago Farina <tfransosi@gmail.com>