sab: reduce the visibility of sqlite statements and their sql queries
Now that they are not prepared in 'ab_init' function and not finilized in
'ab_fini', we can reduce their visibility by restricting them to the function
scope where they are actually used.

As far as it goes for the lifetime, now as local variables within functions,
the variables are initialized each time.

The advantage of having all the pieces for the prepared statement inside each individual
function is great and will help us when doing the Windows port.

Ref:
https://learn.microsoft.com/en-us/cpp/c-language/scope-and-visibility?view=msvc-170
https://learn.microsoft.com/en-us/cpp/c-language/lifetime?view=msvc-170

Signed-off-by: Thiago Farina <tfransosi@gmail.com>