sab: short the lifetime of 'create_sql' variable
Without the 'static' keyword, the lifetime of 'create_sql' variable
will be from the entry into '_create_tables' function to the return from it.
That means it will be initialized everytime '_create_tables' is call. This
is probably going to happen only one time, since 'ab_init' should be called
only once anyway.

With the 'static' storage duration (also called "global") its value value
is defined for the duration of the program. Storage is reserved and the
identifier's stored value is initialized only once, before program startup.

I don't see a reason to keep it defined for the whole duration of the program,
let's just use it and release its memory space when done with it.

Ref:
https://stackoverflow.com/questions/5033627/static-variable-inside-of-a-function-in-c
https://learn.microsoft.com/en-us/cpp/c-language/lifetime?view=msvc-170

Signed-off-by: Thiago Farina <tfransosi@gmail.com>