WunderPeople: move the code for populating the list view into a separate function
The advantage of having it in a separate function is that, the filling of the
list view can be done in later stage and be delayed as long as possible. Right
now we just do it right after creating the list view, but we might find a way
to do it much later.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>