pointers2.c: illustrate the use of dereference operator (*)
The dereference operator is also known as an indirection operator,
which is represented by an asterisk (*).

When using the indirection operator (*) with a pointer variable,
then this operation is known as dereferencing a pointer.

When dereferencing a pointer, the value of the variable pointed by
this pointer will be returned.

And that is the key, because it can be used to access or manipulate the data
stored at the memory location, pointed by the pointer.

https://www.javatpoint.com/c-dereference-pointer

Signed-off-by: Thiago Farina <tfransosi@gmail.com>