vector: declare 'data' member as the last element in the structure
This issue came with the question: How are struct members allocated in memory?

The short answer to this question is: They are allocated in the order as they
were declared in the struct.

The pictorial representation of the 'vector_t' structure memory allocation
for an integer array of 10 elements, for example, would be something like this:

int arr[3] = { 1, 2, 3};

sizeof(int) = 4 bytes
capacity    = 10
size        = 3
data        = &arr

Member name  datasize   capacity   size     data

             ------------------------------------
             |   4    |   10   |    3   | &arr  |
	     |        |        |        |       |
	     ------------------------------------

One reason good reason to have 'datasize', 'capacity' and 'size', all declared
before 'data' member, would be to know all the information about the array,
like the size of its elements, the capacity, and the actual number of elements
the array is holding, before even seeing the array.

It seems to be a good practice to know the required information, before
starting the array manipulation, not the other way.

Without having a dedicated data structure like 'vector_t', you could code
in a non-generic way like this, PROPERTY_ARRAY being a 'vector_t':

typedef struct _RECORD
{
  int id;
  char *name;
} RECORD;

typedef struct _RECORD_ARRAY
{
  size_t cValues;   /* Count of records */
  RECORD *pRecords; /* Array of records */
} RECORD_ARRAY;

Ref:
https://stackoverflow.com/questions/20737176/how-are-struct-members-allocated-in-memory
http://www.fresh2refresh.com/c-programming/c-struct-memory-allocation/
https://stackoverflow.com/questions/5170525/array-in-c-struct
https://stackoverflow.com/questions/36577094/array-of-size-0-at-the-end-of-struct
https://gcc.gnu.org/onlinedocs/gcc/Zero-Length.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>