vector: improve the description of vector_t type
This is to make it clear that it uses an array for its implementation,
and by definition an array is a collection of items stored at contiguous
memory locations.

The fact that it has a dynamic size is only made possible by the vector_t
type itself, because otherwise, arrays have fixed size and are hard to change
after they are initialized.

Refs: Rust std::vec, Java ArrayList

Signed-off-by: Thiago Farina <tfransosi@gmail.com>