bit.c: write power of two numbers in hexadecimal
Silence gcc warnings of "binary constants are a GCC extension".

We use hexadecimal numbers in C because they provide a more friendly
representation of binary-coded values (the 0s and 1s). This happens to be
because each hexadecimal digit represents four bits (binary digits), also
known as a nibble (or nybble). Which means that an 8-bit byte like 11111111
in binary form can be conveniently represented as FF in hexadecimal.

The result of that is that a long sequence of ones and zeros can get turned
into shorter hexadecimal numbers. So hexadecimal numbers are a more compact
way to represent bit sequences.

That explains the reason why we use hexadecimal extensively when programming
in C, rather than writing the sequence of bits directly.

So when in doubt or when writing bit masks or bit flags, use hexadecimal.

Ref:
https://en.wikipedia.org/wiki/Hexadecimal
https://www.ionos.co.uk/digitalguide/server/know-how/hexadecimal-system/

Signed-off-by: Thiago Farina <tfransosi@gmail.com>