tribe: delete contact from the database before deleting from the linked list
Check if deleting from database succeeded. If everything went ok, delete it
from the the linked list.

I think if we do this way, in the future, we will be able to move the linked
list code out of the database code, and make the database code more stateless
leaving to the ui or client code to maintain that linked list.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>