asm/bootloader: append bootloader signature using .word directive
That way it is done in one line.

My machine is Little Endian. I checked with the following code:

+#if (defined __BYTE_ORDER__) && (defined __ORDER_LITTLE_ENDIAN__)
+# if __BYTE_ORDER__ == __ORDER_LITTLE_ENDIAN__
+#  warning "Byte Order is Little Endian"
+# elif __BYTE_ORDER__ == __ORDER_BIG_ENDIAN__
+# endif
+#endif

Since it is little endian, the signature has to be 0xaa55, in this order,
otherwise it does not work. If you invert and try 0x55aa, qemu gives the
error "Boot failed: not a bootable disk".

https://ftp.gnu.org/old-gnu/Manuals/gas-2.9.1/html_chapter/as_7.html
https://github.com/rofl0r/endianness.h/blob/master/endianness.h

Signed-off-by: Thiago Farina <tfransosi@gmail.com>