c-style: document how function names should be written
Since functions are written with the intent to associate them with actions,
they are usually named as a combination of two parts: a transitive verb and a
direct object. In order words, they are usually written in the form of 'verb + noun'.

We will expand and explain more on that later.

Ref:
https://dev.to/somedood/a-grammar-based-naming-convention-13jf
https://learn.microsoft.com/en-us/powershell/scripting/developer/cmdlet/cmdlet-overview?view=powershell-7.3
https://git.sr.ht/~sircmpwn/cstyle#variable-and-function-naming
https://docs.unrealengine.com/5.3/en-US/epic-cplusplus-coding-standard-for-unreal-engine

Signed-off-by: Thiago Farina <tfransosi@gmail.com>