move postfixoperator.c into learn
This is another small program that shows one aspect of how
the postfix increment operator (++) works.

Things get very complicated but interesting when you mix that with
pointers and arrays like *p++. In that little expression you have
many things going on, that we may explain in deep details later.

https://denniskubes.com/2012/08/14/do-you-know-what-p-does-in-c/
https://learn.microsoft.com/en-us/cpp/c-language/c-postfix-increment-and-decrement-operators?view=msvc-170

Signed-off-by: Thiago Farina <tfransosi@gmail.com>