bootloader: use BIOS to print a single character
We can make use of BIOS functions to print a character on screen.
These functions are available to us via the so-called interrupts.

From Ralf Brown's interrupt list we can find that the video interrupt is 0x10.

The function Teletype sounds like a good match, it prints a character given in AL
and automatically advances the cursor.

Then, we can select that function by setting AH to 0x0e, put the ASCII code we want
to print into AL and the call the interrupt 0x10 to actually print it.

Cool.

Ref:
https://medium.com/@g33konaut/writing-an-x86-hello-world-boot-loader-with-assembly-3e4c5bdd96cf

Signed-off-by: Thiago Farina <tfransosi@gmail.com>