c: add a small program to count the number of lines in a file
I built this into a small separate program, to be able to test the
following code more easily and freely:

<code>

 ch = fgetc(fp);
 while (ch != EOF)
 {
     if (ch == '\n')
         numLines++;

     ch = fgetc(fp);
 }

</code>

That part of the code is the one I'm interested, because I'm going to
use it in the numbers program to build an array of that size, that is it,
the size of the number of lines in the file, so the size is not hardcoded.

To test it, is easy, just create a text file, put some numbers or text one per line,
then run `countlines numbers.txt` for example.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>