tribe: return the contact list through an output parameter
We changed the parameter of ab_load_contacts function to a pointer to pointer (
or double pointer) because a double pointer is a technique that with the
unary indirection operator permits to store the address of another pointer
variable and with that we can effectively return the address to the
'contact_list' pointer of type (alpm_list_t) to the caller. And that
eliminates the necessity of calling a second function, ab_get_contact_list,
to actually retrieve it, which was our main goal with this change. It also
had the goal of bringing this API closer to its Windows counterpart (
db.h:GetContactList).

Signed-off-by: Thiago Farina <tfransosi@gmail.com>